+++ dio/lib/src/options.dart
@@ -99,7 +99,6 @@ class BaseOptions extends _RequestConfig with OptionsMixin {
     RequestEncoder? requestEncoder,
     ResponseDecoder? responseDecoder,
     ListFormat? listFormat,
-    this.setRequestContentTypeWhenNoPayload = false,
   })  : assert(connectTimeout == null || !connectTimeout.isNegative),
         assert(baseUrl.isEmpty || Uri.parse(baseUrl).host.isNotEmpty),
         super(
@@ -145,7 +144,6 @@ class BaseOptions extends _RequestConfig with OptionsMixin {
     RequestEncoder? requestEncoder,
     ResponseDecoder? responseDecoder,
     ListFormat? listFormat,
-    bool? setRequestContentTypeWhenNoPayload,
   }) {
     return BaseOptions(
       method: method ?? this.method,
@@ -167,19 +165,8 @@ class BaseOptions extends _RequestConfig with OptionsMixin {
       requestEncoder: requestEncoder ?? this.requestEncoder,
       responseDecoder: responseDecoder ?? this.responseDecoder,
       listFormat: listFormat ?? this.listFormat,
-      setRequestContentTypeWhenNoPayload: setRequestContentTypeWhenNoPayload ??
-          this.setRequestContentTypeWhenNoPayload,
     );
   }
-
-  /// The option will try to imply the default `content-type` header value
-  /// if there is a payload in requests and if the header value not set.
-  ///
-  /// The `content-type` header value will be implied to
-  /// [Headers.jsonContentType] when:
-  /// - [RequestOptions.data] is null and the option is true.
-  /// - [RequestOptions.data] is not null.
-  bool setRequestContentTypeWhenNoPayload;
 }
 
 mixin OptionsMixin {
@@ -310,12 +297,6 @@ class Options {
     if (this.contentType != null) {
       headers[Headers.contentTypeHeader] = this.contentType;
     }
-    // Imply the default content type if capable.
-    // This is a historical issue that dio used to imply the json type, it'll be
-    // too breaking to let user apply a base content type at the moment.
-    if (data != null || baseOpt.setRequestContentTypeWhenNoPayload) {
-      headers[Headers.contentTypeHeader] ??= Headers.jsonContentType;
-    }
     final String? contentType = headers[Headers.contentTypeHeader];
     final extra = Map<String, dynamic>.from(baseOpt.extra);
     if (this.extra != null) {
@@ -393,17 +374,17 @@ class Options {
 
   Duration? _receiveTimeout;
 
-  /// The request Content-Type. The default value is [ContentType.json].
-  /// If you want to encode request body with 'application/x-www-form-urlencoded',
-  /// you can set `ContentType.parse('application/x-www-form-urlencoded')`, and [Dio]
-  /// will automatically encode the request body.
+  /// The request Content-Type.
+  ///
+  /// [Dio] will automatically encode the request body accordingly.
   String? contentType;
 
   /// [responseType] indicates the type of data that the server will respond with
   /// options which defined in [ResponseType] are `json`, `stream`, `plain`.
   ///
-  /// The default value is `json`, dio will parse response string to json object automatically
-  /// when the content-type of response is 'application/json'.
+  /// The default value is [ResponseType.json], [Dio] will parse response string
+  /// to JSON object automatically when the content-type of response is
+  /// [Headers.jsonContentType].
   ///
   /// If you want to receive response data with binary bytes, for example,
   /// downloading a image, use `stream`.
@@ -423,7 +404,8 @@ class Options {
   /// The default value is true
   bool? receiveDataWhenStatusError;
 
-  /// Custom field that you can retrieve it later in [Interceptor]、[Transformer] and the [Response] object.
+  /// Custom field that you can retrieve it later in
+  /// [Interceptor], [Transformer] and the [Response] object.
   Map<String, dynamic>? extra;
 
   /// see [HttpClientRequest.followRedirects],
@@ -718,11 +700,9 @@ class _RequestConfig {
 
   Duration? _receiveTimeout;
 
-  /// The request Content-Type. Defaults to [ContentType.json] if capable.
+  /// The request Content-Type.
   ///
-  /// If you want to encode request body with 'application/x-www-form-urlencoded',
-  /// you can set `ContentType.parse('application/x-www-form-urlencoded')`,
-  /// and [Dio] will automatically encode the request body.
+  /// [Dio] will automatically encode the request body accordingly.
   String? get contentType => _headers[Headers.contentTypeHeader] as String?;
 
   set contentType(String? contentType) {